// frontend/types/factorDetail.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */
import type { DataTypeEnum } from './dataTypeEnum';

/**
 * Serializer completo para detalhes do fator.
 */
export interface FactorDetail {
  readonly id: number;
  /** @maxLength 100 */
  name: string;
  /** @maxLength 20 */
  symbol: string;
  data_type?: DataTypeEnum;
  /**
   * Number of decimal places for quantitative factors
   * @minimum -9223372036854776000
   * @maximum 9223372036854776000
   */
  precision?: number;
  /** JSON with levels configuration */
  levels_config?: unknown;
  readonly experiment: number;
  readonly experiment_title: string;
  /** @pattern ^[-a-zA-Z0-9_]+$ */
  readonly experiment_slug: string;
  readonly created_at: string;
  readonly updated_at: string;
}