// frontend/types/optimizationGoalEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `maximize` - Maximize
* `minimize` - Minimize
* `target` - Target
* `none` - Nenhum(a/as)
 */
export type OptimizationGoalEnum = typeof OptimizationGoalEnum[keyof typeof OptimizationGoalEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OptimizationGoalEnum = {
  maximize: 'maximize',
  minimize: 'minimize',
  target: 'target',
  none: 'none',
} as const;